Major cleanup of internal documentation.
Also cleaned up local variable declarations as much as possible,  hoisting them to the top of the relevant function. Coded incomplete and very rough first-cut at a "replacement" getDestiationData() function. This was suspended, when I realized that changes to "MPMUTILS" were warranted, and more information from Ethan and Drashti is required before this can be completed. Checking in now so the work completed thus far is captured.